KumaNekoBackend/requirements.txt
astroid==2.2.5
autopep8==1.4.4
colorama==0.4.1
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.8.2
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.4
wrapt==1.11.1
